+++ History.md
@@ -6,7 +6,7 @@
   * if file was renamed during upload to avoid conflict,
     new field 'originalName' keeps original name
   * dynamic uploadDir and uploadUrl
-  * upload.getFiles() added, see README.md for example
+  * upload.getFiles() added
 
 0.0.3 / 2012-11-25
 ==================